Add LICENSE.txt (MIT)
I've made multiple failed attempts to contact the author of the original
code about the intended license.

The original code and icons were published without a license.

Since then, most of the code has been rewritten and all of the icons
have been replaced, so it should be OK to add a license.